+++ src/components/FilterByTags/interfaces.ts
@@ -1,4 +1,4 @@
-import { IGenresProps } from '../../types/IGenresProps';
+import { IGenresProps } from '../../core/types/IGenresProps';
 
 export interface IFilterByTagsProps {
   options: IGenresProps;
