+++ src/components/Movie/MovieList/interfaces.ts
@@ -1,4 +1,4 @@
-import { IMovieProps } from '../../../types/IMovieProps';
+import { IMovieProps } from '../../../core/types/IMovieProps';
 
 export interface IMovieListProps {
   movies: IMovieProps[];
