+++ src/components/Search/interfaces.ts
@@ -1,0 +1,3 @@
+export interface ISearchProps {
+  className?: string;
+}
