+++ src/components/UI/DatePickerInput/DatePickerInput.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import cn from 'classnames';
 import DatePicker from 'react-datepicker';
 import { MONTH, YEARS } from '../../../core/constants';
-import { getYear, getMonth } from '../../../utils/date';
+import { getYear, getMonth } from '../../../core/utils/date';
 import { IDatePickerInputProps } from './interfaces';
 import { Icon } from '../Icon';
