+++ src/components/UI/DatePickerInput/ReactDatepickerCustom.css
@@ -1,3 +1,4 @@
+/* stylelint-disable selector-class-pattern */
 .react-datepicker {
   color: #fff;
   background-color: #424242;
