+++ src/core/api/mocked/fetchMovieById.ts
@@ -1,5 +1,5 @@
 import MOVIES from '../../data/movies.json';
-import { IMovieProps } from '../../../types/IMovieProps';
+import { IMovieProps } from '../../types/IMovieProps';
 
 export const fetchMovieById = async (id: string): Promise<IMovieProps> => {
   const movie = MOVIES.find(item => item.id === id);
