+++ src/core/api/mocked/fetchMovies.ts
@@ -1,5 +1,5 @@
 import MOVIES from '../../data/movies.json';
-import { IMovieProps } from '../../../types/IMovieProps';
+import { IMovieProps } from '../../types/IMovieProps';
 
 export const fetchMovies = async (): Promise<IMovieProps[]> => {
   return Promise.resolve(MOVIES);
