+++ src/core/types/IGenresProps.ts
