+++ src/core/types/IMovieProps.ts
